chore: add crud function for action model

+++ src/controllers/action.controller.ts
@@ -1,116 +1,68 @@
 import {
-  Count,
-  CountSchema,
-  Filter,
   repository,
-  Where,
 } from '@loopback/repository';
 import {
   post,
   param,
   get,
-  getFilterSchemaFor,
   getModelSchemaRef,
-  getWhereSchemaFor,
   patch,
-  put,
   del,
   requestBody,
+  HttpErrors,
 } from '@loopback/rest';
-import {Action} from '../models';
-import {ActionRepository} from '../repositories';
+import { Action } from '../models';
+import { ActionRepository } from '../repositories';
+import { ActionRequestBody } from './specs/action-controller.specs';
+import { authenticate, AuthenticationBindings, UserProfile } from '@loopback/authentication';
+import { inject } from '@loopback/core';
 
 export class ActionController {
   constructor(
     @repository(ActionRepository)
-    public actionRepository : ActionRepository,
-  ) {}
+    public actionRepository: ActionRepository,
+  ) { }
 
   @post('/actions', {
     responses: {
       '200': {
         description: 'Action model instance',
-        content: {'application/json': {schema: getModelSchemaRef(Action)}},
+        content: { 'application/json': { schema: getModelSchemaRef(Action) } },
       },
     },
   })
+  @authenticate('jwt')
   async create(
-    @requestBody({
-      content: {
-        'application/json': {
-          schema: getModelSchemaRef(Action, {exclude: ['id']}),
-        },
-      },
-    })
+    @inject(AuthenticationBindings.CURRENT_USER)
+    currentUserProfile: UserProfile,
+    @requestBody(ActionRequestBody)
     action: Omit<Action, 'id'>,
   ): Promise<Action> {
-    return this.actionRepository.create(action);
+    const { id } = currentUserProfile;
+    const newAction = { ...action, userId: id };
+    return this.actionRepository.create(newAction);
   }
 
-  @get('/actions/count', {
+  @get('/actions/{status}', {
     responses: {
       '200': {
-        description: 'Action model count',
-        content: {'application/json': {schema: CountSchema}},
-      },
-    },
-  })
-  async count(
-    @param.query.object('where', getWhereSchemaFor(Action)) where?: Where<Action>,
-  ): Promise<Count> {
-    return this.actionRepository.count(where);
-  }
-
-  @get('/actions', {
-    responses: {
-      '200': {
-        description: 'Array of Action model instances',
+        description: 'Array of Action model instances by status',
         content: {
           'application/json': {
-            schema: {type: 'array', items: getModelSchemaRef(Action)},
+            schema: { type: 'array', items: getModelSchemaRef(Action) },
           },
         },
       },
     },
   })
-  async find(
-    @param.query.object('filter', getFilterSchemaFor(Action)) filter?: Filter<Action>,
+  @authenticate('jwt')
+  async getActionsByStatus(
+    @param.query.string('status') status: string,
+    @inject(AuthenticationBindings.CURRENT_USER)
+    currentUserProfile: UserProfile,
   ): Promise<Action[]> {
-    return this.actionRepository.find(filter);
-  }
-
-  @patch('/actions', {
-    responses: {
-      '200': {
-        description: 'Action PATCH success count',
-        content: {'application/json': {schema: CountSchema}},
-      },
-    },
-  })
-  async updateAll(
-    @requestBody({
-      content: {
-        'application/json': {
-          schema: getModelSchemaRef(Action, {partial: true}),
-        },
-      },
-    })
-    action: Action,
-    @param.query.object('where', getWhereSchemaFor(Action)) where?: Where<Action>,
-  ): Promise<Count> {
-    return this.actionRepository.updateAll(action, where);
-  }
-
-  @get('/actions/{id}', {
-    responses: {
-      '200': {
-        description: 'Action model instance',
-        content: {'application/json': {schema: getModelSchemaRef(Action)}},
-      },
-    },
-  })
-  async findById(@param.path.string('id') id: string): Promise<Action> {
-    return this.actionRepository.findById(id);
+    const { id } = currentUserProfile;
+    return this.actionRepository.find({ where: { userId: id, status } });
   }
 
   @patch('/actions/{id}', {
@@ -120,32 +72,28 @@ export class ActionController {
       },
     },
   })
+  @authenticate('jwt')
   async updateById(
     @param.path.string('id') id: string,
+    @inject(AuthenticationBindings.CURRENT_USER)
+    currentUserProfile: UserProfile,
     @requestBody({
       content: {
         'application/json': {
-          schema: getModelSchemaRef(Action, {partial: true}),
+          schema: getModelSchemaRef(Action, { partial: true }),
         },
       },
     })
     action: Action,
   ): Promise<void> {
+    const userId = currentUserProfile.id;
+    const shouldUpdateAction = await this.actionRepository.findOne({
+      where: { userId, id },
+    });
+    if (!shouldUpdateAction) {
+      throw new HttpErrors.BadRequest('Action isn\'t existed or isn\'t owned by yourself!');
+    }
     await this.actionRepository.updateById(id, action);
-  }
-
-  @put('/actions/{id}', {
-    responses: {
-      '204': {
-        description: 'Action PUT success',
-      },
-    },
-  })
-  async replaceById(
-    @param.path.string('id') id: string,
-    @requestBody() action: Action,
-  ): Promise<void> {
-    await this.actionRepository.replaceById(id, action);
   }
 
   @del('/actions/{id}', {
@@ -155,7 +103,19 @@ export class ActionController {
       },
     },
   })
-  async deleteById(@param.path.string('id') id: string): Promise<void> {
+  @authenticate('jwt')
+  async deleteById(
+    @param.path.string('id') id: string,
+    @inject(AuthenticationBindings.CURRENT_USER)
+    currentUserProfile: UserProfile,
+  ): Promise<void> {
+    const userId = currentUserProfile.id;
+    const shouldUpdateAction = await this.actionRepository.findOne({
+      where: { userId, id },
+    });
+    if (!shouldUpdateAction) {
+      throw new HttpErrors.BadRequest('Action isn\'t existed or isn\'t owned by yourself!');
+    }
     await this.actionRepository.deleteById(id);
   }
 }
